extracts editor pane and creates a toEnum for ease of adding tabs
